+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Brilliant trait to chaos gold
+### Changed
+- Chaos rune and enchantment token recipes are now created with data gens (should have no changes)
+
 ## [3.7.5] - 2020-12-06
 ### Added
 - Configs to adjust the weights of or completely disable generation of specific gem ores [#491]